CW2/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Marshmallow==0.14.0
pyodbc==4.0.32
